Moved helper files to util folder and broke up each helper method into seperate files
